Extraction de RuleValueVignette de Results.js
