+++ package.json
@@ -62,7 +62,7 @@
     "css-loader": "^0.28.1",
     "csv-loader": "^2.1.1",
     "daggy": "^1.1.0",
-    "eslint": "^4.4.1",
+    "eslint": "^4.8.0",
     "eslint-plugin-react": "^7.0.1",
     "express": "^4.15.3",
     "fantasy-combinators": "0.0.1",
