+++ source/components/Results.js
@@ -4,16 +4,12 @@ import classNames from 'classnames'
 import {Link} from 'react-router-dom'
 import {connect} from 'react-redux'
 import { withRouter } from 'react-router'
-import {formValueSelector} from 'redux-form'
 
 import './Results.css'
-import {capitalise0} from '../utils'
-import {computeRuleValue, clearDict} from 'Engine/traverse'
+import {clearDict} from 'Engine/traverse'
 import {encodeRuleName} from 'Engine/rules'
 import {getObjectives} from 'Engine/generateQuestions'
-
-let fmt = new Intl.NumberFormat('fr-FR').format
-let humanFigure = decimalDigits => value => fmt(value.toFixed(decimalDigits))
+import RuleValueVignette from './RuleValueVignette'
 
 @withRouter
 @connect(
@@ -49,54 +45,17 @@ export default class Results extends Component {
 							<i className="fa fa-arrow-circle-left" aria-hidden="true"></i>Reprendre la simulation
 						</Link>
 					</div>
-				: <div id="results-titles">
+					: <div id="results-titles">
 						<h2>Vos résultats <i className="fa fa-hand-o-right" aria-hidden="true"></i></h2>
 						{do {let text = R.path(['simulateur', 'résultats'])(analysedSituation.root)
 							text &&
 								<p id="resultText">{text}</p>
 						}}
 						<p id="understandTip"><i className="fa fa-lightbulb-o" aria-hidden="true"></i><em>Cliquez pour comprendre chaque calcul</em></p>
-				</div>
+					</div>
 				}
 				<ul>
-					{explanation.map(
-						({name, dottedName, type, 'non applicable si': nonApplicable, formule: {nodeValue: formuleValue}}) =>
-						do {
-							//TODO quel bordel, à revoir
-							let
-								ruleValue = computeRuleValue(formuleValue, nonApplicable && nonApplicable.nodeValue),
-								unsatisfied = ruleValue == null,
-								nonApplicableValue = nonApplicable ? nonApplicable.nodeValue : false,
-								irrelevant = nonApplicableValue === true || formuleValue == 0,
-								number = nonApplicableValue == false && formuleValue != null
-
-								;<li key={name} className={classNames({unsatisfied, irrelevant, number})}>
-
-
-								<Link to={"/regle/" + encodeRuleName(name)} >
-									<div className="rule-type">
-										{type}
-									</div>
-									<div className="rule-box">
-
-										<div className="rule-name">
-											{capitalise0(name)}
-										</div>
-										<p>
-										{conversationStarted && (
-											irrelevant ?
-												"Vous n'êtes pas concerné"
-												: unsatisfied ?
-													'En attente de vos réponses...'
-													: <span className="figure">{humanFigure(2)(formuleValue) + '€'}</span>
-										)}
-										</p>
-									</div>
-								</Link>
-								{/* <div className="pointer">•</div> */}
-							</li>
-						}
-					)}
+					{explanation.map( rule => <RuleValueVignette key={rule.nom} {...rule} conversationStarted={conversationStarted} />)}
 				</ul>
 			</section>
 		)
